Removed support for strategies. All versions are saved as new documents.
Supporting strategies made the module more complicated than it should be and some of the features existed only in the Array strategy making them incompatible. Saving all versions in an array only works for reasonable small datasets as there is a maximum of data a document can contain. That limit will bring issues on large number of saved versions acting as a time bomb on a project.